- more 64bit support

+++ libs/SalesforceSDKCore/SalesforceSDKCore/Classes/Security/SFAbstractPasscodeViewController.m
@@ -143,7 +143,7 @@ NSUInteger const kMaxNumberofAttempts = 10;
             action = SFSecurityLockoutActionPasscodeVerified;
             break;
         default:
-            [self log:SFLogLevelError format:@"Unknown passcode controller mode: %d.  No security lockout action will be configured.", self.mode];
+            [self log:SFLogLevelError format:@"Unknown passcode controller mode: %lu.  No security lockout action will be configured.", (unsigned long)self.mode];
             action = SFSecurityLockoutActionNone;
             break;
     }
